Code/WindowsApp/requirements.txt
pyserial
pyqt5
pyqtgraph